request changes refactor
